doc(ngx-oauth2-access-token): standard types
