+++ projects/mrpachara/ngx-oauth2-access-token/src/lib/types/standard/README.md
@@ -1,0 +1,8 @@
+### References
+- [JSON Web Token (JWT)](https://www.rfc-editor.org/rfc/rfc7519.html)
+- [JSON Web Signature (JWS)](https://www.rfc-editor.org/rfc/rfc7515.html)
+- [JSON Web Algorithms (JWA)](https://www.rfc-editor.org/rfc/rfc7518.html)
+- [OpenID Connect Core 1.0](https://openid.net/specs/openid-connect-core-1_0.html)
+- [JSON Web Key (JWK)](https://www.rfc-editor.org/rfc/rfc7517.html)
+- [The OAuth 2.0 Authorization Framework](https://www.rfc-editor.org/rfc/rfc6749.html)
+- [Proof Key for Code Exchange by OAuth Public Clients](https://www.rfc-editor.org/rfc/rfc7636.html)
