+++ projects/mrpachara/ngx-oauth2-access-token/src/lib/types/standard/jwt/token.types.ts
@@ -1,0 +1,1 @@
+export type JwtTokenType = `${string}.${string}`;
